refactor(kitchenpos): add JPA, Spring annotations & replace AS-IS code
