+++ src/main/java/kitchenpos/common/domain/Price.java
@@ -3,8 +3,17 @@ package kitchenpos.common.domain;
 import java.math.BigDecimal;
 import java.util.Objects;
 
+import javax.persistence.Column;
+import javax.persistence.Embeddable;
+
+@Embeddable
 public class Price {
-	private final BigDecimal value;
+	@Column(name = "price")
+	private BigDecimal value;
+
+	protected Price() {
+
+	}
 
 	public Price(Long value) {
 		if (value == null || !(value >= 0)) {
