+++ src/main/java/kitchenpos/products/domain/Product.java
@@ -1,49 +1,50 @@
 package kitchenpos.products.domain;
 
+import java.util.UUID;
+
 import javax.persistence.Column;
+import javax.persistence.Embedded;
 import javax.persistence.Entity;
 import javax.persistence.Id;
 import javax.persistence.Table;
-import java.math.BigDecimal;
-import java.util.UUID;
+
+import kitchenpos.common.domain.DisplayedName;
+import kitchenpos.common.domain.Price;
 
 @Table(name = "product")
 @Entity
 public class Product {
-    @Column(name = "id", columnDefinition = "varbinary(16)")
-    @Id
-    private UUID id;
+	@Column(name = "id", columnDefinition = "varbinary(16)")
+	@Id
+	private UUID id;
+	@Embedded
+	private DisplayedName name;
+	@Embedded
+	private Price price;
 
-    @Column(name = "name", nullable = false)
-    private String name;
+	protected Product() {
 
-    @Column(name = "price", nullable = false)
-    private BigDecimal price;
+	}
 
-    public Product() {
-    }
+	public Product(DisplayedName name, Price price) {
+		this.id = UUID.randomUUID();
+		this.name = name;
+		this.price = price;
+	}
 
-    public UUID getId() {
-        return id;
-    }
+	public void changePrice(Price price) {
+		this.price = price;
+	}
 
-    public void setId(final UUID id) {
-        this.id = id;
-    }
+	public UUID getId() {
+		return id;
+	}
 
-    public String getName() {
-        return name;
-    }
+	public DisplayedName getName() {
+		return name;
+	}
 
-    public void setName(final String name) {
-        this.name = name;
-    }
-
-    public BigDecimal getPrice() {
-        return price;
-    }
-
-    public void setPrice(final BigDecimal price) {
-        this.price = price;
-    }
+	public Price getPrice() {
+		return price;
+	}
 }
