--- src/main/java/kitchenpos/products/tobe/domain/ProductRepository.java
@@ -1,16 +1,0 @@
-package kitchenpos.products.tobe.domain;
-
-import java.util.List;
-import java.util.Optional;
-import java.util.UUID;
-
-public interface ProductRepository {
-	Product save(Product product);
-
-	Optional<Product> findById(UUID id);
-
-	List<Product> findAll();
-
-	List<Product> findAllByIdIn(List<UUID> ids);
-}
-
