+++ src/main/java/kitchenpos/products/ui/ProductRestController.java
@@ -9,6 +9,8 @@ import java.net.URI;
 import java.util.List;
 import java.util.UUID;
 
+import javax.validation.Valid;
+
 @RequestMapping("/api/products")
 @RestController
 public class ProductRestController {
@@ -19,14 +21,19 @@ public class ProductRestController {
     }
 
     @PostMapping
-    public ResponseEntity<Product> create(@RequestBody final Product request) {
+    public ResponseEntity<Product> create(
+        @Valid @RequestBody final ProductCreateRequest request
+    ) {
         final Product response = productService.create(request);
         return ResponseEntity.created(URI.create("/api/products/" + response.getId()))
             .body(response);
     }
 
     @PutMapping("/{productId}/price")
-    public ResponseEntity<Product> changePrice(@PathVariable final UUID productId, @RequestBody final Product request) {
+    public ResponseEntity<Product> changePrice(
+        @PathVariable final UUID productId,
+        @Valid @RequestBody final ProductPriceChangeRequest request
+    ) {
         return ResponseEntity.ok(productService.changePrice(productId, request));
     }
