+++ src/test/java/kitchenpos/menus/application/MenuServiceTest.java
@@ -5,7 +5,7 @@ import kitchenpos.menus.domain.MenuGroupRepository;
 import kitchenpos.menus.domain.MenuProduct;
 import kitchenpos.menus.domain.MenuRepository;
 import kitchenpos.common.infra.FakePurgomalumClient;
-import kitchenpos.products.application.InMemoryProductRepository;
+import kitchenpos.products.infra.InMemoryProductRepository;
 import kitchenpos.products.domain.Product;
 import kitchenpos.products.domain.ProductRepository;
 import kitchenpos.common.domain.PurgomalumClient;
