+++ src/test/java/kitchenpos/products/application/ProductServiceTest.java
@@ -1,131 +1,121 @@
 package kitchenpos.products.application;
 
-import kitchenpos.common.infra.FakePurgomalumClient;
-import kitchenpos.menus.application.InMemoryMenuRepository;
-import kitchenpos.menus.domain.Menu;
-import kitchenpos.menus.domain.MenuRepository;
-import kitchenpos.products.domain.Product;
-import kitchenpos.products.domain.ProductRepository;
-import kitchenpos.common.domain.PurgomalumClient;
+import static org.assertj.core.api.Assertions.*;
+import static org.junit.jupiter.api.Assertions.*;
+
+import java.util.List;
+import java.util.UUID;
+
 import org.junit.jupiter.api.BeforeEach;
-import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.NullSource;
 import org.junit.jupiter.params.provider.ValueSource;
 
-import java.math.BigDecimal;
-import java.util.List;
-import java.util.UUID;
+import kitchenpos.Fixtures;
+import kitchenpos.common.domain.DisplayedName;
+import kitchenpos.common.domain.Price;
+import kitchenpos.common.domain.PurgomalumClient;
+import kitchenpos.common.infra.FakePurgomalumClient;
+import kitchenpos.products.domain.Product;
+import kitchenpos.products.domain.ProductRepository;
+import kitchenpos.products.infra.InMemoryProductRepository;
+import kitchenpos.products.ui.ProductCreateRequest;
+import kitchenpos.products.ui.ProductPriceChangeRequest;
 
-import static kitchenpos.Fixtures.*;
-import static org.assertj.core.api.Assertions.assertThat;
-import static org.assertj.core.api.Assertions.assertThatThrownBy;
-import static org.junit.jupiter.api.Assertions.assertAll;
+public class ProductServiceTest {
+	private ProductRepository productRepository;
+	private PurgomalumClient purgomalumClient;
+	private ProductService productService;
 
-class ProductServiceTest {
-    private ProductRepository productRepository;
-    private MenuRepository menuRepository;
-    private PurgomalumClient purgomalumClient;
-    private ProductService productService;
+	@BeforeEach
+	void setUp() {
+		productRepository = new InMemoryProductRepository();
+		purgomalumClient = new FakePurgomalumClient();
+		productService = new ProductService(productRepository, purgomalumClient);
+	}
 
-    @BeforeEach
-    void setUp() {
-        productRepository = new InMemoryProductRepository();
-        menuRepository = new InMemoryMenuRepository();
-        purgomalumClient = new FakePurgomalumClient();
-        productService = new ProductService(productRepository, menuRepository, purgomalumClient);
-    }
+	@Test
+	void 상품을_등록할_수_있다() {
+		// given
+		ProductCreateRequest request = new ProductCreateRequest("후라이드", 16_000L);
 
-    @DisplayName("상품을 등록할 수 있다.")
-    @Test
-    void create() {
-        final Product expected = createProductRequest("후라이드", 16_000L);
-        final Product actual = productService.create(expected);
-        assertThat(actual).isNotNull();
-        assertAll(
-            () -> assertThat(actual.getId()).isNotNull(),
-            () -> assertThat(actual.getName()).isEqualTo(expected.getName()),
-            () -> assertThat(actual.getPrice()).isEqualTo(expected.getPrice())
-        );
-    }
+		// when
+		Product actual = productService.create(request);
 
-    @DisplayName("상품의 가격이 올바르지 않으면 등록할 수 없다.")
-    @ValueSource(strings = "-1000")
-    @NullSource
-    @ParameterizedTest
-    void create(final BigDecimal price) {
-        final Product expected = createProductRequest("후라이드", price);
-        assertThatThrownBy(() -> productService.create(expected))
-            .isInstanceOf(IllegalArgumentException.class);
-    }
+		// then
+		assertThat(actual).isNotNull();
+		assertAll(
+			() -> assertThat(actual.getId()).isNotNull(),
+			() -> assertThat(actual.getName()).isEqualTo(new DisplayedName(request.getName(), purgomalumClient)),
+			() -> assertThat(actual.getPrice()).isEqualTo(new Price(request.getPrice()))
+		);
+	}
 
-    @DisplayName("상품의 이름이 올바르지 않으면 등록할 수 없다.")
-    @ValueSource(strings = {"비속어", "욕설이 포함된 이름"})
-    @NullSource
-    @ParameterizedTest
-    void create(final String name) {
-        final Product expected = createProductRequest(name, 16_000L);
-        assertThatThrownBy(() -> productService.create(expected))
-            .isInstanceOf(IllegalArgumentException.class);
-    }
+	@ParameterizedTest
+	@ValueSource(longs = -1_000L)
+	@NullSource
+	void 상품의_가격이_올바르지_않으면_등록할_수_없다(Long price) {
+		// given
+		ProductCreateRequest request = new ProductCreateRequest("후라이드", price);
 
-    @DisplayName("상품의 가격을 변경할 수 있다.")
-    @Test
-    void changePrice() {
-        final UUID productId = productRepository.save(product("후라이드", 16_000L)).getId();
-        final Product expected = changePriceRequest(15_000L);
-        final Product actual = productService.changePrice(productId, expected);
-        assertThat(actual.getPrice()).isEqualTo(expected.getPrice());
-    }
+		// when & then
+		assertThatThrownBy(() -> productService.create(request))
+			.isInstanceOf(IllegalArgumentException.class);
+	}
 
-    @DisplayName("상품의 가격이 올바르지 않으면 변경할 수 없다.")
-    @ValueSource(strings = "-1000")
-    @NullSource
-    @ParameterizedTest
-    void changePrice(final BigDecimal price) {
-        final UUID productId = productRepository.save(product("후라이드", 16_000L)).getId();
-        final Product expected = changePriceRequest(price);
-        assertThatThrownBy(() -> productService.changePrice(productId, expected))
-            .isInstanceOf(IllegalArgumentException.class);
-    }
+	@ParameterizedTest
+	@ValueSource(strings = {"비속어", "욕설이 포함된 이름"})
+	@NullSource
+	void 상품의_이름이_올바르지_않으면_등록할_수_없다(String name) {
+		// given
+		ProductCreateRequest request = new ProductCreateRequest(name, 16_000L);
 
-    @DisplayName("상품의 가격이 변경될 때 메뉴의 가격이 메뉴에 속한 상품 금액의 합보다 크면 메뉴가 숨겨진다.")
-    @Test
-    void changePriceInMenu() {
-        final Product product = productRepository.save(product("후라이드", 16_000L));
-        final Menu menu = menuRepository.save(menu(19_000L, true, menuProduct(product, 2L)));
-        productService.changePrice(product.getId(), changePriceRequest(8_000L));
-        assertThat(menuRepository.findById(menu.getId()).get().isDisplayed()).isFalse();
-    }
+		// when & then
+		assertThatThrownBy(() -> productService.create(request))
+			.isInstanceOf(IllegalArgumentException.class);
+	}
 
-    @DisplayName("상품의 목록을 조회할 수 있다.")
-    @Test
-    void findAll() {
-        productRepository.save(product("후라이드", 16_000L));
-        productRepository.save(product("양념치킨", 16_000L));
-        final List<Product> actual = productService.findAll();
-        assertThat(actual).hasSize(2);
-    }
+	@Test
+	void 상품의_가격을_변경할_수_있다() {
+		// given
+		Product given = Fixtures.product("후라이드", 16_000L);
+		UUID givenId = productRepository.save(given).getId();
+		ProductPriceChangeRequest request = new ProductPriceChangeRequest(15_000L);
 
-    private Product createProductRequest(final String name, final long price) {
-        return createProductRequest(name, BigDecimal.valueOf(price));
-    }
+		// when
+		Product actual = productService.changePrice(givenId, request);
 
-    private Product createProductRequest(final String name, final BigDecimal price) {
-        final Product product = new Product();
-        product.setName(name);
-        product.setPrice(price);
-        return product;
-    }
+		// then
+		assertThat(actual.getPrice()).isEqualTo(new Price(request.getPrice()));
+	}
 
-    private Product changePriceRequest(final long price) {
-        return changePriceRequest(BigDecimal.valueOf(price));
-    }
+	@ParameterizedTest
+	@ValueSource(longs = -1_000L)
+	@NullSource
+	void 상품의_가격이_올바르지_않으면_변경할_수_없다(Long price) {
+		// given
+		Product given = Fixtures.product("후라이드", 16_000L);
+		UUID givenId = productRepository.save(given).getId();
+		ProductPriceChangeRequest request = new ProductPriceChangeRequest(price);
 
-    private Product changePriceRequest(final BigDecimal price) {
-        final Product product = new Product();
-        product.setPrice(price);
-        return product;
-    }
+		// when & then
+		assertThatThrownBy(() -> productService.changePrice(givenId, request))
+			.isInstanceOf(IllegalArgumentException.class);
+	}
+
+	// TODO : 상품의 가격이 변경될 때 메뉴의 가격이 메뉴에 속한 상품 금액의 합보다 크면 메뉴가 숨겨진다.
+
+	@Test
+	void 상품의_목록을_조회할_수_있다() {
+		// given
+		productRepository.save(Fixtures.product("후라이드", 16_000L));
+		productRepository.save(Fixtures.product("양념치킨", 16_000L));
+
+		// when
+		List<Product> actual = productService.findAll();
+
+		// then
+		assertThat(actual).hasSize(2);
+	}
 }
