+++ src/test/java/kitchenpos/products/infra/InMemoryProductRepository.java
@@ -1,4 +1,4 @@
-package kitchenpos.products.tobe.application;
+package kitchenpos.products.infra;
 
 import java.util.ArrayList;
 import java.util.HashMap;
@@ -8,8 +8,8 @@ import java.util.Optional;
 import java.util.UUID;
 import java.util.stream.Collectors;
 
-import kitchenpos.products.tobe.domain.Product;
-import kitchenpos.products.tobe.domain.ProductRepository;
+import kitchenpos.products.domain.Product;
+import kitchenpos.products.domain.ProductRepository;
 
 public class InMemoryProductRepository implements ProductRepository {
 	private final Map<UUID, Product> products = new HashMap<>();
